requirements.txt
# -----------------------------
# 🚀 CORE BACKEND
# -----------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0

# -----------------------------
# 🤖 GEMINI (CORRECT SDK)
# -----------------------------
google-generativeai==0.3.2

# -----------------------------
# 📦 DATA VALIDATION
# -----------------------------
pydantic==2.5.0

# -----------------------------
# ⚙️ ENV MANAGEMENT
# -----------------------------
python-dotenv==1.0.0

# -----------------------------
# 🌐 HTTP CLIENT (for testing & async calls)
# -----------------------------
httpx==0.25.1

# -----------------------------
# 🧪 TESTING
# -----------------------------
pytest==7.4.3
pytest-asyncio==0.21.1

# -----------------------------
# 🖼️ TEMPLATING
# -----------------------------
jinja2==3.1.2